+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ7FDGHYXNPTVB0VC5Q8J27M"
+        "01KJ83AQR39BSR2Y86WD2NYM6A"
     )
     r"""The unique identifier of the data source"""