+++ packages/orq-rc/RELEASES.md
@@ -6059,3 +6059,13 @@ Based on:
 - [python v4.5.0-rc.0] packages/orq-rc
 ### Releases
 - [PyPI v4.5.0-rc.0] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.0 - packages/orq-rc
+
+## 2026-02-24 15:14:51
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.727.0 (2.839.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.1] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.1] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.1 - packages/orq-rc